Disable `decoderEnforceMaxRstFramesPerWindow` for HTTP/2 clients
Motivation:

Netty 4.1.100 introduced a new feature for HTTP/2 to protect from
DDoS vector, and we added system properties to control its settings in
#2728. In version 4.1.101 Netty disabled this protection for clients
by default, but due to how we use our system properties, that change
won't automatically apply to ServiceTalk.

Modifications:

- `OptimizedHttp2FrameCodecBuilder.decoderEnforceMaxRstFramesPerWindow`
now considers `isServer` flag to decide if the inferred values should
be applied or not;

Result:

`decoderEnforceMaxRstFramesPerWindow` is enabled only for HTTP/2
servers.