loadbalancer: Change generics of LoadBalancerPolicy (#2773)
Motivation:

The generic types that represent the resolved address and the
real type of the connection are generics to the _method_ of
LoadBalancingPolicy.buildSelector. This means the the policy
must support all types of addresses and connections making it
more difficult to specialize.

Modifications:

Move the generics from the buildSelector method to the
LoadBalancerPolicy type itself.