concurrent-api: fix IllegalArgumentExcetion for zero jitter in RetryStrategies
Motivation:

If you pass in a jitter of 0 for the retry strategies you'll get
an IllegalArgumentException thrown by the `Random.nextLong(lower, upper)`
method because the lower and upper are the same value. This is because
the upper bound argument is non-inclusive.

Modifications:

Add 1 to each call to extend the range by 1 nanosecond and make the full
delay range selectable.

Result:

What is the result of this change?