loadbalancer: Simplify ConnectionFactory usage in DefaultHost (#2796)
Motivation:

We currently use the ConnectionFactory in two places in DefaultHost:
- creating new connections
- health checking
Each of them has to go through the trouble of trying to add the
newly created connection to the connection pool and making
sure the state remains coherent.

Modifications:

DefaultHosts health checking feature now uses the `Host.newConnection`
method instead of using the factory directly. This lets it
get the connection management benefit already in
`DefaultHost.newConnection`. It also makes the path to extracting
the health check a bit easier to follow.

Result:

What is the result of this change?