loadbalancer: Add the ErrorClass enum and use it in RequestTracker (#2808)
Motivation:

RequestTracker has three observer methods: `onSuccess`, `onCancel`,
and `onError`. onCancel is really a special case of `onError` but it
can be helpful to distinguish them at times which leads to the more
general problem: it is often helpful to know what kind of error
it was so we can establish the origin.

Modifications:

- Add the `ErrorClass` enum for classifying `RequestTracker` errors.
- Consolidate the `onCancel` method into the `onError` method by
  having a `CANCELLED` variant of the `ErrorClass` enum.

Result:

Less methods and more information about the origin of errors.