loadbalancer: remove 'linear search space' notion from DefaultLoadBalancer
Motivation:

The configurable `linearSearchSpace` is used as a strategy
for picking among sessions in a particular host but it is not
highly used and mayber for good reason: sessions that are beyond
the max size are always in the odd position of getting picked a
lot or seldom, depending on how many sessions there are. There is
not way to set how many times it will attempt to pick a random
entry after the linear search space so if you exceed the linear
space by just 1 the extra connection will be re-attempted 64
times in a row before failing.
There is probably room for customizing the connection pool
selection strategy but we can probably make it more modular.

Modifications:

Delete the notion of a `linearSearchSpace` from the DefaultLoadBalancer.

Result:

Less code to worry about.