loadbalancer: Pass connection failures to the HealthIndicator
Motivation:

We can't rely on the request path to pass in connection
failures because if we can't get a connection we can never
add the observers. That means we need the `Host` implementation
to pass it in directly.

Modifications:

Add two new methods to `HealthIndicator` that provide the current
time and and a callback for submitting connection failures.

Result:

We should be able to track connection failures now and factor
those into the `HealthTracker` implementations.