loadbalancer: move DefaultLoadBalancer to its own experimental module
Motivation:

We want to start integrating with other packages but we can't
without making some interfaces and types public. However, that
means we're committing to the API and we don't want to do that
yet.

Modifications:

Move the new load balancer code to its own experimental module.
This is a strong signal to users that the API is in flux and
should not be dependend upon yet.