http-netty: Properly set the host header for MultiAddressUrlClient
Motivation:

Users of the MultiAddressHttpClientBuilder are expected to put in
a request target that includes the authority, which may or may not
include the port. The authority should be directly reflected in the
`Host` header but right now we will add both host and port
regardless of what was specified.

Modifications:

Copy the authority information directly from the request when
overrides haven't been specified.