Support for Optional TLS negotiation for http/1 and http/2
Motivation
----------
In certain deployment scenarios it can be useful to accept
incoming traffic on the same port for both TLS and non-TLS
connections. ServiceTalk does not support this at this
point, so this changeset aims to add the feature.

Modifications
-------------

User-facing, a new option is introduced on the ServerSslConfig
called "acceptInsecureConnections". When set to true, http/1
and http/2 connections are accepted both encrypted and unencrypted.

This still works with SNI and ALPN, since the negotiation happens
when the connection is first accepted and the first bytes are checked
to see if a SSL request is made or not.

Internally, the OptionalSSLHandler will do the check and afterwards
the OptionalSslNegotiator will add the corresponding handlers depending
on the protocol and tls/non-tls to the pipeline. At this point the other
code takes over.

Tests have been added to validate the feature and avoid regressions.