Add support for Optional TLS negotiation for http/1 and http/2 (#2871)
Motivation
----------
In certain deployment scenarios it can be useful to accept incoming traffic
on the same port for both TLS and non-TLS connections. ServiceTalk does not
support this at this point, so this changeset aims to add the feature.

Modifications
-------------
User-facing, a new overload option for the sslConfig is introduced which allows
the user to specify that insecure connections should also be accepted. When
set to true, both http/1 and http/2 protocols are supported, although ServiceTalk
does not support h2c cleartext upgrades so this combination is still not supported.

This also works with ALPN and SNI, since the negotiation happens before the
TLS connection is fully established.

Internally, the OptionalSslNegotiator will determine if the incoming first bytes
of the connection signify a TLS connection or not and will hand it to the
right channel initializer afterwards.