Optimize protocol negotiation with Optional TLS
Motivation
----------
The initial support for Optional TLS did only allow insecure connections
if either H1 or H2 is enabled, but not protocols at the same time. This
is a limitation when used together with ALPN on secure connections and
can be improved.

Modifications
-------------
This changeset makes sure that ALPN support still works when optional
TLS is configured. In the case where both H1 and H2 are enabled on
the server side and a insecure connection has been negotiated, the
code now falls back to H1 support only. This means that h2c cleartext
upgrades are not supported, but they haven't been supported inside
ServiceTalk anyways.

Result
------
Best possible ALPN support at the moment in combination with
Optional TLS.