loadbalancer-experimental: fix broken assert in XdsOutlierDetector
Motivation:

When cancelling the whole xDS outlier detector it first enters the
sequential executor then has each child indicator cancel itself. It
does this synchronously and then checks that the indicator set is
empty. This doesn't work like that because each indicator tries to
bounce itself through the sequential executor, so it won't actually
be empty until after the method exits.

Modifications:

- Use the `sequentialCancel()` method instead: we know that we're
  in the sequential executor already.
- Add some tests.

Result:

What is the result of this change?