loadbalancer-experimental: make connection selection strategy modular (#2815)
Motivation:

The configurable linearSearchSpace is used as a strategy
for picking among sessions in a particular host but it is not
highly used and mayber for good reason: sessions that are beyond
the max size are always in the odd position of getting picked a
lot or seldom, depending on how many sessions there are. There is
not way to set how many times it will attempt to pick a random
entry after the linear search space so if you exceed the linear
space by just 1 the extra connection will be re-attempted 64
times in a row before failing. There are potentially better ways to
select hosts such as encouraging a core pool of connections, etc.

Modifications:

Make the connection selection policy modular and have multiple
supported patterns that users can experiment with.