`RetryingHttpRequesterFilter`: add on request retry callback (#2916)
Motivation:

There are use-cases when users need to perform an action before the
retry, like updating meta-data or logging/metrics. Because all retry
functions return `BackoffPolicy`, it's still not known inside the
function of `BackoffPolicy` will retry or not because of the retry
counts limit.

Modifications:
- Add `RetryingHttpRequesterFilter.Builder.onRequestRetry(RetryCallbacks)`
that users can use to intercept every retry attempt;
- Test that the new callback works for request retries;

Result:

Users can see when the retry actually happens, after backoff time.

If there will be a similar use-case for `reserveConnection` in the
future, we will add a separate callback.