loadbalancer-experimental: break test dependency cycle
Motivation:

IJ can get upset that there is a dependency cycle between http-netty
and loadbalancer-experimental-provider because the provider module
depends on http-netty as a test dependency.

Modifications:

Move the tests from loadbalancer-experimental-provider to htp-netty
and remove the test dependency. This is a better test structure
anyway because if people have http-netty on the classpath it should
load the provider.

Result:

Better test structure and fix IJ tools.