Convert scheme to lower case when parsing URI (#2938)
Motivation:

`HttpRequestMetaData#scheme()` javadocs state the result will be lowercase, but this is not always the case because `Uri3986` parses the scheme as-is. As a result, `DefaultMultiAddressUrlHttpClientBuilder` may have more than 1 client instance for the same address if users use URI with different scheme letters case. 
See https://github.com/apple/servicetalk/issues/1423 for details.

Modifications:

- Convert scheme to lower case when parsing `Uri3986` reusing well-known schemes ("http" and "https")

Result:

Fixes #1423.