loadbalancer-experimental: make loadBalancingWeight easier to debug
Motivation

Because we use mutable state to represent the load balancing weight
we need to make sure that it is reset back to the SD weight before
we re-run hosts through the transformations. It's currently done when
we set the SD weight which happens even if it is the same, so we can
change the lb weight even if we don't plan to run it through the
transformations again. We will still load balance correctly because
we won't rebuild, but it could be really confusing to debug.

Modifications

- Separate setting the SD weight and LB weight and before we rebuild
  reset the LB weight to the current SD weight.
- Add some tests that ensure we set the weight correctly.