loadbalancer-experimental: make loadBalancingWeight easier to debug (#2954)
Motivation

Because we use mutable state to represent the load balancing weight
we need to make sure that it is reset back to the service discovery weight
before we re-run hosts through the transformations. This is currently done
when we set the SD weight via `serviceDiscoveryWeight(double)` which
happens even if it is unchanged. The result is that we change the lb weight
even if we don't plan to run it through the transformations again. We still
load balance correctly because we won't rebuild, but it could be really
confusing to debug since the lb weight won't reflect transformations.

Modifications

- Separate setting the SD weight and LB weight and before we rebuild
  reset the LB weight to the current SD weight.
- Add some tests that ensure we set the weight correctly.