loadbalancer-experimental: Add lbDescription to host priority strategy
Motivation:

Logs are going to be exponentially more useful if we know which LB
they're tied to but right now the DefaultHostPriorityStrategy is flying
blind with regard to it's parent.

Modifications:

Follow the pattern of other components of DefaultLoadBalancer and pass
the lbDescription into the construction of the HostPriorityStrategy.

Result:

Better observability.