FromInputStreamPublisher: avoid extra allocation of a buffer
Motivation:

In #2949 we optimized a case when `available()` is not implemented and
always returns `0`. However, we de-optimized a use-case when it's
implemented because after that change the last call to `available()`
always returns 0, but we still allocate a buffer of size `readChunkSize`
that won't be used at all.

Modifications:
- Enhance `doNotFailOnInputStreamWithBrokenAvailableCall(...)` test
before any changes to have better test coverage.
- Remove `byte[] buffer` from a class variable. It can be a local
variable because it's never reused in practice. Only the last `buffer`
won't be used nullified, but we don't need it after that.
- When `available()` returns `0`, try reading a single byte and then
check availability again instead of always falling back to
`readChunkSize`.
- Adjust `doNotFailOnInputStreamWithBrokenAvailableCall()` test to
account for the 2nd call to `available()`;
- Add `singleReadTriggersMoreAvailability()` test to simulate when the
2nd call to `available()` returns positive value;

Result:

1. No allocation of a `buffer` that won't be used at the EOF.
2. Account for new availability if it appears after a `read()`.