`DefaultHostAndPort`: equals and hashCode are not consistent
Motivation:

If two instances of `DefaultHostAndPort` are created with hostname in
different case, they pass `equals` check but don't have a consistent
`hashCode`. As a result, users can't query a `Map` entry with keys in
different case.
Side-effect of that is a multi-address http client that may have
multiple instances of a single-address client for the same hostname in
different cases, preventing it from reusing connections.

Modifications:
- Implement `DefaultHostAndPort.hashCode()` in a consistent way with
`InetSocketAddressHolder` where `hostName` hash is computed in
case-insensitive mode.
- Re-generate `DefaultHostAndPort.equals(...)` implementation to allow a
check by reference.
- Enhance `DefaultMultiAddressUrlHttpClientBuilderTest` to demonstrate
that the problem went away.

Result:

1. `DefaultHostAndPort` can be used as a case-insensitive key for maps,
matching definition of the host name in URL.
2. Multi-address client correctly recognizes the same host name in
case-insensitive way.