loadbalancer-experimental: allow configuring the pending request penalty
Motivation:

The pending request penalty is a way to try to avoid connections with
lost of pending requests which results in a more even request distribution.
However, the penalty is currently fixed which doesn't allow us to choose
how to prioritize between the fastest hosts vs the most fair request
distribution.

Modifications:

Make the penalty factor configurable.