Set-Cookie names must not be lowercase
This changeset aligns the "Set-Cookie" propertie names with the
RFC6265 in the sense that according to syntax (see
https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-rfc6265bis#section-4.1.1)
all the names must uppercase the first character.

Most browsers are lax in how they interpret the cookies so the
old behavior should not cause any issues, but its better to
align with the spec.