Align Set-Cookie attribute names with RFC6265 syntax (#2995)
This changeset aligns the "Set-Cookie" property names with the RFC6265 in the sense that according to the spec (see https://datatracker.ietf.org/doc/html/draft-ietf-httpbis-rfc6265bis#section-4.1.1) where property names are using camel-case syntax instead of all lower case.

NOTE: The name of an attribute-value pair is not case sensitive. So while they are presented here in CamelCase, such as "HttpOnly" or "SameSite", any case is accepted. E.x.: "httponly", "Httponly", "hTTPoNLY", etc.

This means that this change is purely cosmetic and helps with consistency with other implementations, i.e. Netty.