Fix discovery flow for `PartitionedHttpClient`
Motivation:

Our load balancers have an ability to cancel and re-subscribe to the
`ServiceDiscovery` publisher as an attempt to trigger re-discovery when
LB turns into unhealthy state for all addresses. However, for
`PartitionedHttpClient` it will result in `DuplicateSubscribeException`
because `groupToMany` operator does not support re-subscribes.

Modifications:

- Apply `multicast` operator with `cancelUpstream=false` flag to allow
LB of a single-address client to re-subscribe to the same group
publisher;
- Disable a retry strategy for `ServiceDiscovery` at single-address
client level for partitioned client;
- Implement `NoRetriesStrategy.toString()` for better logging;
- Reproduce described scenario in `PublisherGroupToManyTest`;

Result:

Single-address clients under partitioned client won't fail with
`DuplicateSubscribeException` if LB decides to re-subscribe to the
group's stream.