LoadBalancer should always consider the first events as initial state
Motivation:

Currently, both implementations of the `LoadBalancer` assume 2
possibilities for the first events collection after re-subscribe:
"state of the world" or "delta" and tries to infer which one it is based
on what event statuses are present in the collection. However, it
doesn't provide any guarantees because even if SD impl keeps returning
deltas after a re-subscribe, a delta can contain only `AVAILABLE`
events, but it will be incorrectly interpreted as "state of the world".

Because of Reactive Streams rule clarified in #3002, `LoadBalancer`
implementations should always expect "state of the world" for the
first events after re-subscribe.

Modifications:

- Remove inference of the first collection of events after resubscribe
and always expect a new subscriber to start from "state of the world".
- Update `resubscribeToEventsWhenAllHostsAreUnhealthy()` test to reflect
behavior change.

Result:

Both `LoadBalancer` implementations follow `ServiceDiscoverer` contract
and expect the first collection of events after re-subscribe to
represent a "state of the world".