transport-netty-internal: Finish reading when input shutdown
Motivation:

Netty 4.2 changed the behavior of its epoll implementation to
and that resulted in not receiving the ChannelInputShutdownReadComplete
event. This is because we don't have a read interest set at the
time because it's an idle client channel and we now won't get
the event until we attempt to read.

:servicetalk-transport-netty-internal:spotbugsMain

Modifications:

Watch for the ChannelInputShutdownEvent and if we get it
trigger a read. This will cause the event loop to drain any
remaining bytes from the channel and emit the event.

Result:

4.2 build should pass more tests.