loadbalancer-experimental: reduce dogpiling on hosts after healthy hosts
Motivation:

Round-robin algorithms very often suffer from bugs where the host
after a healthy host will tend to get more traffic. This is because
the algorithm will then try to advance to the 'next' index.

Modifications:

If we find a bad host attempt to advance the cursor so the next
selection doesn't also pick it.

Result:

Less dogpiling behavior.