capacity-limiter-api: avoid division in EMA
Motivation:

Division is much more expensive than multiplication, certainly when
the compliler doesn't know the denominator and can do tricks to mitigate.

Modifications:

In the capacity limiter EMA we have a division operation where the
denominator is always the same which is a good candidate for inverting
once and turning into a multiplication operation.