concurrent-api: Add ability to cleanup resources on timeout
Motivation:

The Single.timeout(..) family of operations are part of a class of
operations will short-circuit the response based on the cancellation
pathway, specifically they will return a TimeoutException down the
error path. Because cancellation and the result are intrinsically
racy there is currently the possibility that the result will be
generated but we cannot return it to the subscriber having already
given them an Exception in the error pathway. For values that are
stateful this can result in a resource leak.

Modifications:

- Add a timeout combinator that includes a resource cleanup hook.
- Use that hook in the AbstractTimeoutHttpFilter where we need to
  drain the message body which can include live and stateful
  resources such as cleanup hooks and even the life cycle management
  of a HTTP session.

Result:

One less leak and at least the option for users to cleanup stateful
resources on timeout races.