http-utils: fix leak in BeforeFinallyHttpOperator (#3034)
Motivation:

BeforeFinallyHttpOperator has the option to short-circuit and not
emit a response if it has been cancelled. However, the response has
a (potentially) stateful body that needs to be cleaned up and this
isn't done if we happen to get an onSuccess after cancellation (which
is an intrinsic race in RS).

Modifications:

- Always cleanup the response body if we receive it after a
cancellation.

Result:

One less response body leak.