traffic-resilience-http: Add debug logging for leaked tickets
Motivation:

It can be really tough to know what happened to a capacity limiter
ticket if it gets abandoned.

Modifications:

Add an optional debug mode that can help. It will add a finalizer
wrapper that can tell if the ticket was abandoned. If it was, it
logs and clean up.