concurrent-api: save the timestamp of the SingleToFuture.get() calls (#3051)
Motivation:

It can be very difficult to know if a thread is slow or completely
stuck without taking multiple thread dumps and comparing them.

Modifications:

Try to mitigate this by saving the time stamp of the last blocking
`.get()` call. This should then show up in heap dumps and we can
at least compare them to see if they've all started blocking at
about the same time or if there is a big spread. The latter would
suggest they are truly stuck.