loadbalancer-experimental: make DefaultLoadBalancer the default RR implementation
Motivation:

It's been out a while now and DefaultLoadBalancer can exactly behave as
the old RoundRobinLoadBalancer. That means we can insert it in place of
RoundRobinLoadBalancer.

Modifications:

Default to building a DefaultLoadBalancer in round-robin compatibility mode
via the RoundRobinToDefaultLBMigrationProvider. We retain the flag for safety
for a few releases but will eventually remove it and RoundRobinLoadBalancer.

Result:

What is the result of this change?